[#25] Rename package to meet naming convention (see https://rules.sonarsource.com/java/RSPEC-120)
