+++ DakoChatApplication/.idea/runConfigurations/UdpServer.xml
@@ -1,6 +1,6 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="UdpServer" type="Application" factoryName="Application" folderName="AuditLogServerNew" singleton="false" nameIsGenerated="true">
-    <option name="MAIN_CLASS_NAME" value="edu.hm.dako.chat.AuditLogServerNew.UdpServer" />
+    <option name="MAIN_CLASS_NAME" value="edu.hm.dako.chat.audit_log_server.UdpServer" />
     <module name="DakoChatApplication" />
     <extension name="coverage">
       <pattern>
