+++ DakoChatApplication/src/main/java/edu/hm/dako/chat/audit_log_server/AuditLogModel.java
@@ -1,4 +1,4 @@
-package edu.hm.dako.chat.AuditLogServerNew;
+package edu.hm.dako.chat.audit_log_server;
 
 public class AuditLogModel extends AbstractAuditLogModel {
 }
