+++ DakoChatApplication/src/main/java/edu/hm/dako/chat/audit_log_server/ServerInterface.java
@@ -1,4 +1,4 @@
-package edu.hm.dako.chat.AuditLogServerNew;
+package edu.hm.dako.chat.audit_log_server;
 
 public interface ServerInterface {
     /**
